Move examples to their own module.
