+++ kubernetes-examples/src/main/java/io/fabric8/kubernetes/examples/WatchExample.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.client.examples;
+package io.fabric8.kubernetes.examples;
 
 import io.fabric8.kubernetes.api.model.ReplicationController;
 import io.fabric8.kubernetes.client.Config;
